adding testing for utility
